requirements.txt
discord.py>=2.3.2
beautifulsoup4>=4.12.2
pytz>=2024.1
python-dotenv>=1.0.1
requests>=2.31.0